requirements.txt
att-event-engine
att_service_manager
pymongo
grpcio
Flask
Flask-Session